Fix use of requirements.txt in github workflows
Since requirements.txt is removed, we need to use the
pyproject.toml/poetry route to install packages. Poetry doesn't work as
easily inside the docker containers - using the virtual environment
poetry creates requires extra setup and making sure paths are correct,
etc. Telling poetry not to create the virtual environment can break
things when the packages are installed due to version conflicts between
what poetry wants and what the product wants. The workaround is to have
poetry create a requirements type file and then use the normal pip
install to install the packages, not using poetry or the virtual
environment after that.